changed all components names to index, and component.styles.ts to styles.ts
