+++ src/components/ArtWorkCard/styles.ts
