+++ src/components/ArtWorkDetails/styles.ts
