+++ src/screens/HomeScreen.tsx
@@ -2,8 +2,8 @@ import React, {useCallback} from 'react';
 import {SafeAreaView, Linking} from 'react-native';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import {HOME_SCREEN} from '../navigation/constants';
-import Catalog from '../components/Catalog/Catalog';
-import HomeHeader from '../components/HomeHeader/HomeHeader';
+import Catalog from '../components/Catalog';
+import HomeHeader from '../components/HomeHeader';
 import {artInstitute, getMapCoords} from '../utils/methods';
 import {RootBottomParamList} from '../navigation/BottomTabNavigator';
